+++ packages/adminui-pub/CHANGELOG.md
@@ -6,8 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - new method `UBApp.downloadDocument` - Download a document from BLOB store directly into a file without
+   loading it into memory as with `getDocument`. Can be used to download a huge documents what
+   do not fit in the browser memory.  
 
 ### Changed
+ - `Download` action of BasePanel `Document` menu uses new method `UBApp.downloadDocument` to save a document
+   to the file system without loading it into browser memory.
 
 ### Deprecated